+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 
 0.2.7 (YYYY-MM-DD)
 ------------------
+* Ignore row dimension when fixing column shapes (:pr:`165`)
 * Fix zarr coordinate writes (:pr:`162`)
 * Deprecate Python 3.6 (:pr:`161`)
 * Add IMAGING_WEIGHT_SPECTRUM to default Measurement Schema (:pr:`160`)